config 'mode': removed 'all', added 'orb'; renamed module 'nrb' to 'ard'
