+++ config.ini
@@ -9,11 +9,11 @@
 
 [PROCESSING]
 
-# OPTIONS: sar | ard | all
+# OPTIONS: sar | nrb | orb
 # sar: only SAR processing
-# ard: only ARD generation from existing SAR products
-# all: SAR processing and Analysis Ready Data (ARD) generation
-mode = all
+# nrb: NRB generation from existing SAR products
+# orb: ORB generation from existing SAR products
+mode = sar, nrb
 
 # optional definition of a geometry via tile IDs or a vector file.
 # [aoi_tiles] expects a comma-separated list of MGRS tile IDs.
